% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Plastics Data}
\usage{
load_data()
}
\value{
A dataframe containing the plastics data from tidytuesday.
}
\description{
Load Plastics Data
}
\examples{
data <- load_data()
}
